+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Release date: `2024-12-xx`
 - Added the printer constant `DEFAULT_HEAT_TIME` (integer, defaults to `80`).
 - Added the printer constant `DEFAULT_HEAT_INTERVAL` (integer, defaults to `12`).
 - Added the printer constant `DEFAULT_MOST_HEATED_POINT` (integer, defaults to `3`).
+- Added the convenient `ThermalPrinter.has_paper` property.
 - Added the `ThermalPrinter.__exit__()` method to properly close the printer when leaving the context manager.
 
 # 0.3.0